refactor(get): added extra return var for operation data
The get function now returns some info about the record in the operation result such as version and consumed capacity.
