+++ result.go
@@ -2,8 +2,8 @@ package dynastorev2
 
 import "github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 
-// MutationResult returned with update and create to provide some information about the update
-type MutationResult struct {
+// OperationResult returned with operations to provide some information about the update
+type OperationResult struct {
 	Version          int64
 	ConsumedCapacity *types.ConsumedCapacity
 }
